+++ CHANGELOG.md
@@ -29,7 +29,7 @@ for now removed features.
 
 - **select**: internal value was not changed when the select is closed and when we used keyboard nav on it (key up, 
 key down... to change the current selected option)
-- **doc**: improve `wcs-select` documentation
+- **doc**: improve `wcs-select` and `wcs-input` documentation
 - **modal**: (accessibility) focus modal controls which served to open the modal on closure
 
 ### Security